requirements.txt
streamlit>=1.24.0
snowflake-snowpark-python>=1.11.0
pandas>=1.5.0
python-dotenv>=1.0.0
tabulate>=0.9.0
